[Implementation] Optimization
Optimized a lot of things, mainly pausing processing when sand is not moving. Implemented velocity and acceleration. Program now works with small sand sizes without major lag as before.

+++ PixelPhysics/core/src/com/aidenx11/game/elements/Element.java
@@ -123,7 +123,7 @@ public abstract class Element {
 		this.modified = modified;
 	}
 
-	public int getVelocity() {
+	public float getVelocity() {
 		return velocity;
 	}
